theta-back/pyproject.toml
[project]
name = "theta-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.40.0",
]
